issue-168 src/component/classes/adapter/ + tokens
